+++ RELEASES.md
@@ -6,6 +6,7 @@ Released on TBD (UTC).
 - Added optional beta weighting and percent option greeks (#2317), thanks @faysou
 - Added precision inference for `TardisCSVDataLoader`, where `price_precision` and `size_precision` are now optional
 - Added `UnixNanos::to_datetime_utc()` in Rust
+- Added `Mark` variant for `PriceType` enum
 
 ### Breaking Changes
 - Removed [talib](https://github.com/nautechsystems/nautilus_trader/tree/develop/nautilus_trader/indicators/ta_lib) subpackage (see deprecations for v1.211.0)